Initial commit
Clean-room rewrite of my emacs.d. I've been using the same emacs
configuration since high school and it had gotten a bit hairy over
the years. Turns out, a good deal of it didn't do a thing.

So! Here we are with a clean-room implementation. I've only got Rust
set up at the moment as that's what I'm working with the majority of
the time nowadays.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>